refactor(cars): move getWinners logic from game class to cars class
